Used size_t for the loop iteration ints

+++ 0x06-pointers_arrays_strings/104-print_buffer.c
@@ -11,7 +11,7 @@
 */
 void print_buffer(char *b, int size)
 {
-	int a, x;
+	size_t a, x;
 	char c;
 
 	if (size <= 0)
